requirements.txt
pyTelegramBotAPI==4.29.1
python-dotenv==1.0.1
Flask==3.1.0
psycopg[binary]==3.2.6